Move webview activity from exam to core
