+++ core/src/main/AndroidManifest.xml
@@ -19,6 +19,11 @@
         <activity
             android:name=".ui.DiscussionActivity"
             android:theme="@style/TestpressTheme" />
+
+        <activity
+            android:name=".ui.WebViewActivity"
+            android:configChanges="orientation|keyboardHidden|screenSize"
+            android:label="Web View" />
     </application>
 
 </manifest>
