+++ course/src/main/java/in/testpress/course/ui/WebViewWithSSO.kt
@@ -4,7 +4,7 @@ import `in`.testpress.core.TestpressCallback
 import `in`.testpress.core.TestpressException
 import `in`.testpress.core.TestpressSdk
 import `in`.testpress.course.R
-import `in`.testpress.exam.ui.WebViewActivity
+import `in`.testpress.ui.WebViewActivity
 import `in`.testpress.fragments.EmptyViewFragment
 import `in`.testpress.fragments.EmptyViewListener
 import `in`.testpress.models.InstituteSettings
